pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "sup-res-bot"
version = "0.1.0"
description = "Discord bot for support/resistance breakout alerts"
authors = [{ name = "SupRes Bot" }]
requires-python = ">=3.10"
dependencies = [
    "discord.py>=2.4.0",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.1"
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
